server/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.36.0
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.109.0
filelock==3.20.1
fsspec==2025.12.0
GeoAlchemy2==0.14.3
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
packaging==25.0
pillow==10.3.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.3
rsa==4.9.1
safetensors==0.7.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.25
shapely==2.1.2
starlette==0.35.1
sympy==1.14.0
tenacity==9.1.2
scikit-learn==1.4.0
tokenizers==0.20.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
